Use double for angles, not int.
The new cargo-intake doesn't travel to a horizontal position, so we've
lost one of our two positions used to extrapolate the conversion
settings from analog values to degrees.

As such, when I calibrated the cargo-intake setpoints to use values
between [0, 1], behold my surprise when I found the PID setpoints stuck
at zero after spending the better part of an hour debugging.

STATUS: Builds.  Tested on testbot.

- Justin C

+++ src/main/cpp/subsystems/CargoIntake.cpp
@@ -126,7 +126,7 @@ void CargoIntake::RotateToPosition(wpi::StringRef configName) {
     RotateToPosition(ang);
 }
 
-void CargoIntake::RotateToPosition(int worldAngle) {
+void CargoIntake::RotateToPosition(double worldAngle) {
     m_RotationPID.SetSetpoint(worldAngleToMachine(worldAngle));
     m_RotationPID.Enable();
 }
